forgot password functionality, send email

+++ controllers/bootcamps.js
@@ -53,6 +53,16 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route   POST /api/v1/bootcamps
 // @access  Private
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
+    req.body.user = req.user.id;
+
+    const createdBootcamp = await Bootcamp.findOne({ user: req.user.id });
+
+    if (createdBootcamp && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`User has already published a bootcamp`, 400)
+        );
+    };
+
     const data = await Bootcamp.create(req.body);
 
     res.status(201).json({success: true, data});
@@ -62,16 +72,24 @@ exports.createBootcamp = asyncHandler(async (req, res, next) => {
 // @route   PUT /api/v1/bootcamps/:id
 // @access  Private
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-    const data = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-        new: true,
-        runValidators: true,
-    })
+    const bootcamp = await Bootcamp.findById(req.params.id);
 
-    if (!data) {
+    if (!bootcamp) {
         return next(
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
         );  
-    }
+    };
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`Not authorized to update the bootcamp`, 401)
+        );
+    };
+
+    const data = await bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+        new: true,
+        runValidators: true,
+    });
 
     res.status(200).json({success: true, data});
 });
@@ -86,6 +104,12 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
         return next(
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
         );  
+    };
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !==  'admin') {
+        return next(
+            new ErrorResponse('Not authorized to delete the bootcamp', 401)
+        )
     };
 
     await bootcamp.deleteOne();
@@ -104,6 +128,12 @@ exports.uploadBootcampImage = asyncHandler(async (req, res, next) => {
         return next(
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`, 404)
         );  
+    };
+
+    if (bootcamp.user.toString() !== req.user.id && req.user.role !== 'admin') {
+        return next(
+            new ErrorResponse(`Not authorized to upload image to the bootcamp`, 401)
+        );
     };
 
     if (!req.files) {
